% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_tab.R
\name{dat_tab}
\alias{dat_tab}
\title{Data table for cluster size - total}
\usage{
dat_tab(strat, dataset)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}

\item{dataset}{custom dataset or subset (in case of multiple estimates from same data)}
}
\description{
Generates a data table with an overview of the number of observations,
the distance from the transect line to each observation, and the tetraonid cluster size.
}
\examples{
dat_tab(strat = LEVEL)
dat_tab(dataset = DATA, strat = LEVEL)
}
\keyword{cluster}
\keyword{size}
\keyword{table}
\keyword{total}
