% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_tab_ad.R
\name{dat_tab_ad}
\alias{dat_tab_ad}
\title{Data table for cluster size - adults only}
\usage{
dat_tab_ad(strat, dataset)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}

\item{dataset}{custom dataset or subset (in case of multiple estimates from same data)}
}
\description{
Generates a data table with an overview of the number of observations,
the distance from the transect line to each observation, and the adult tetraonid cluster size.
}
\examples{
dat_tab_ad(strat = LEVEL)
dat_tab_ad(dataset = DATA, strat = LEVEL)
}
\keyword{adult}
\keyword{cluster}
\keyword{size}
\keyword{table}
