% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Desc_all.R
\name{Desc_all}
\alias{Desc_all}
\title{Data table for descriptive statistics - total}
\usage{
Desc_all()
}
\description{
Displays summarized statistics, including no. of transect lines and kilometers
taxated, number of total observations and the number of birds counted.
}
\examples{
Desc_all
}
\keyword{descriptive}
\keyword{statistics}
\keyword{table}
\keyword{total}
