% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Desc_omr.R
\name{Desc_omr}
\alias{Desc_omr}
\title{Data table for descriptive statistics - per area}
\usage{
Desc_omr()
}
\description{
Displays summarized statistics per area, including number of transect lines and 
kilometers inventoried, number of total observations and the number of birds counted.
}
\examples{
Desc_omr
}
\keyword{area}
\keyword{descriptive}
\keyword{statistics}
\keyword{table}
