% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felles_res.R
\name{felles_res}
\alias{felles_res}
\title{Data table for model estimates - combined total and per area}
\usage{
felles_res(strat = strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'Omradenavn' or 'Year'}
}
\description{
Displays the estimated total density of tetraonids, the density of
adults, and the production of juveniles, with upper and lower
confidence levels and coefficient of variation. Additionally, displays 
the estimates for each area ('OmradeNavn'), if this level
of stratification is specified.
}
\examples{
felles_res(strat = LEVEL)
}
\keyword{combined}
\keyword{estimates}
\keyword{model}
\keyword{table}
