% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig1.R
\name{Fig1}
\alias{Fig1}
\title{Figure - detection probability}
\usage{
Fig1(klasser = 11, model1 = ds.model1)
}
\arguments{
\item{klasser}{Number of classes in the histogram}

\item{model1}{The Distance sampling model to be used}
}
\description{
Plots two graphs: a) detection probability with increasing distance from 
the transect line, and b) detection curve with standard error.
}
\examples{
\dontrun{
Fig1(klasser=11, model1=ds.model1)
}
}
\keyword{detection}
\keyword{figure}
\keyword{probability}
