% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omr_res.R
\name{omr_res}
\alias{omr_res}
\title{Data table for model estimates - per area}
\usage{
omr_res(strat = strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}
}
\description{
Displays the estimated total density of tetraonids, the density of
adults, and the production of juveniles, with upper and lower
confidence levels and coefficient of variation, for each level of
stratification ('OmradeNavn').
NB! 'strat' must be 'OmradeNavn' - only use if this level of stratification
has been used throughout the analysis.
}
\examples{
omr_res(strat = LEVEL)
}
\keyword{area}
\keyword{estimates}
\keyword{model}
\keyword{table}
