% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec.R
\name{rec}
\alias{rec}
\title{Estimate brood size}
\usage{
rec(strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}
}
\description{
Estimates brood size, i.e. the number of chicks per adult pair,
based on the level of stratification.
}
\examples{
rec(strat = LEVEL)
}
\keyword{brood}
\keyword{size}
