% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_skogsfugl.R
\name{rec_skogsfugl}
\alias{rec_skogsfugl}
\title{Estimate brood size - forest birds}
\usage{
rec_skogsfugl(strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}
}
\description{
Estimates brood size, i.e. the number of chicks per adult female,
based on the level of stratification.
NB! Only for use with forest birds, i.e. Capercaillie and Black grouse.
}
\examples{
rec(strat = LEVEL)
}
\keyword{birds}
\keyword{brood}
\keyword{forest}
\keyword{size}
