% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_tab.R
\name{reg_tab}
\alias{reg_tab}
\title{Data table for stratification level area coverage}
\usage{
reg_tab(strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}
}
\description{
Generates a data table with an overview of the area covered for each
stratification level ('No'/'Year' defaults to 1000).
}
\examples{
reg_tab(strat = LEVEL)
}
\keyword{area}
\keyword{coverage}
\keyword{table}
