% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Res_FeFo.R
\name{Res_FeFo}
\alias{Res_FeFo}
\title{Data table for model estimates - FeFo
Displays the estimated total density of tetraonids, the density of
adults, and the production of juveniles, with upper and lower
confidence levels and coefficient of variation. This function only
applies to FeFo (Finnmarkseiendommen), and includes total estimates
and estimates per region.}
\usage{
Res_FeFo(strat = strat)
}
\arguments{
\item{strat}{Level of stratification ('No', 'OmradeNavn' or 'Year')}
}
\description{
Data table for model estimates - FeFo
Displays the estimated total density of tetraonids, the density of
adults, and the production of juveniles, with upper and lower
confidence levels and coefficient of variation. This function only
applies to FeFo (Finnmarkseiendommen), and includes total estimates
and estimates per region.
}
\examples{
\dontrun{
Res_FeFo(strat = LEVEL)
}
}
\keyword{FeFo}
\keyword{estimates}
\keyword{model}
\keyword{table}
