% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resultat_for_opplast_OmrID.R
\name{Resultat_for_opplast_OmrID}
\alias{Resultat_for_opplast_OmrID}
\title{Data table for export of estimates - per area}
\usage{
Resultat_for_opplast_OmrID(userID = 16, Navn = "Erlend Nilsen")
}
\arguments{
\item{userID}{The user ID of the person analysing the data}

\item{Navn}{The name of the person analysing the data}
}
\description{
Generates a data table with estimates, confidence levels, brood sizes
etc. per area, ready for import into HFP.
}
\examples{
\dontrun{
Resultat_for_opplast_OmrID(userID=16, Navn="Erlend Nilsen")
}
}
\keyword{area}
\keyword{export}
\keyword{table}
