% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resultat_for_opplast_RappNiva_skogsfugl.R
\name{Resultat_for_opplast_RappNiva_skogsfugl}
\alias{Resultat_for_opplast_RappNiva_skogsfugl}
\title{Data table for export of estimates - forest birds}
\usage{
Resultat_for_opplast_RappNiva_skogsfugl(userID = as.integer(),
  Navn = as.character())
}
\arguments{
\item{userID}{The user ID of the person analysing the data}

\item{Navn}{The name of the person analysing the data}
}
\description{
Generates a data table with estimates, confidence levels, brood sizes
etc., ready for import into HFP. Only applies to forest birds, i.e.
Capercaillie and Black grouse.
}
\examples{
Resultat_for_opplast_RappNiva_skogsfugl(userID=16, Navn="Erlend Nilsen")
}
\keyword{birds}
\keyword{export}
\keyword{forest}
\keyword{table}
