% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samle_res.R
\name{samle_res}
\alias{samle_res}
\title{Data table for model estimates - total}
\usage{
samle_res(strat = strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}
}
\description{
Displays the estimated total density of tetraonids, the density of
adults, and the production of juveniles, with upper and lower
confidence levels and coefficient of variation.
}
\examples{
samle_res(strat = LEVEL)
}
\keyword{estimates}
\keyword{model}
\keyword{table}
\keyword{total}
