% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_tab.R
\name{samp_tab}
\alias{samp_tab}
\title{Data table for sampling effort}
\usage{
samp_tab(strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}
}
\description{
Generates a data table with an overview of the LineID's, stratification
level ('År' or 'Områdenavn') and the effort (distance taxated) for each LineID.
}
\examples{
samp_tab(strat = LEVEL)
}
\keyword{effort}
\keyword{sampling}
\keyword{table}
