% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skogsfugl_res.R
\name{skogsfugl_res}
\alias{skogsfugl_res}
\title{Data table for model estimates - forest birds only}
\usage{
skogsfugl_res(strat = strat)
}
\arguments{
\item{strat}{Level of stratification - 'No', 'OmradeNavn' or 'Year'}
}
\description{
Displays the estimated total density of tetraonids, the density of
adults, and the production of juveniles, with upper and lower
confidence levels and coefficient of variation. For use with
Capercaillie and Black grouse only.
}
\examples{
skogsfugl_res(strat = LEVEL)
}
\keyword{birds}
\keyword{estimates}
\keyword{forest}
\keyword{model}
\keyword{table}
